add: lua get_scene_count function
