+++ src/filters/lua/lua_scene_api.hpp
@@ -1,5 +1,5 @@
 //
-// Created by leondietrich on 2/11/26.
+// Created by leondietrich on 2/11/25.
 // SPDX-License-Identifier: GPL-3.0-or-later
 //
 #pragma once
@@ -7,5 +7,5 @@
 #include <sol/sol.hpp>
 
 namespace dmxfish::filters::lua {
-    void init_lua_scene_api(sol::state& lua);
+    void init_lua_color_api(sol::state& lua);
 }
